requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg[binary]==3.2.3
psycopg-pool==3.2.4
cryptography==43.0.3
argon2-cffi==23.1.0
httpx==0.27.2
typer==0.12.5
rich==13.9.4
pytest==8.3.3
